requirements.txt
boto3
requests
python-telegram-bot
luis
